pyproject.toml
[project]
name = "clearmind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.2",
    "google-adk>=1.27.3",
    "google-genai",
    "python-dotenv"
]
